test: move all the tests for new features into a separate package
